+++ cmd/run.go
@@ -12,11 +12,21 @@ import (
 
 var flagRunPath string
 var flagStage string
+var flagEnv string
 
 var runCmd = &cobra.Command{
 	Use:   "run",
-	Short: "A brief description of your command",
-	Long: `A longer description that spans multiple lines and likely contains examples
+	Short: "Runs a new build from a .wharf-ci.yml file",
+	Long: `Runs a new build in a Kubernetes cluster using pods
+based on a .wharf-ci.yml file.
+
+If no stage is specified via --stage then wharf-cmd will run all stages
+in sequence, based on their order of declaration in the .wharf-ci.yml file.
+
+All steps in each stage will be run in parallel for each stage.
+
+Read more about the .wharf-ci.yml file here:
+https://iver-wharf.github.io/#/usage-wharfyml/
 `,
 	RunE: func(cmd *cobra.Command, args []string) error {
 		stepRun, err := worker.NewK8sStepRunner("build", Kubeconfig)
@@ -25,7 +35,9 @@ var runCmd = &cobra.Command{
 		}
 		stageRun := worker.NewStageRunner(stepRun)
 		b := worker.New(stageRun)
-		def, errs := wharfyml.ParseFile(flagRunPath)
+		def, errs := wharfyml.ParseFile(flagRunPath, wharfyml.Args{
+			Env: flagEnv,
+		})
 		if len(errs) > 0 {
 			log.Warn().WithInt("errors", len(errs)).Message("Cannot run build due to parsing errors.")
 			for _, err := range errs {
@@ -61,4 +73,5 @@ func init() {
 
 	runCmd.Flags().StringVarP(&flagRunPath, "path", "p", ".wharf-ci.yml", "Path to .wharf-ci file")
 	runCmd.Flags().StringVarP(&flagStage, "stage", "s", "", "Stage to run (will run all stages if unset)")
+	runCmd.Flags().StringVarP(&flagEnv, "environment", "e", "", "Environment selection")
 }